Move getToken and common API functions to CommonApi.js
