+++ src/components/InvitePage/Api.js
@@ -1,5 +1,5 @@
 import {BASE_URL} from "../../settings";
-import { getToken } from '../../api/AuthApi.js';
+import { getToken } from '../../api/CommonApi.js';
 
 export function handleTeamCreation(inviteCode, setErrorState, setSuccessState) {
