+++ src/components/QuestPage/Api.js
@@ -1,5 +1,5 @@
 import {BASE_URL} from "../../settings";
-import { getToken } from '../../api/AuthApi.js';
+import { getToken } from '../../api/CommonApi.js';
 
 export default function handleTeamCreation(teamName, questId, setErrorState, setSuccessState) {
     let query = {
